moved checkout functions to other place bcs it makes more sense
