--- api/src/checkout/address/address.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AddressController } from './address.controller';
-
-describe('AddressController', () => {
-  let controller: AddressController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [AddressController],
-    }).compile();
-
-    controller = module.get<AddressController>(AddressController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
