+++ api/src/checkout/checkout.service.ts
@@ -1,10 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CheckoutViewmodel } from './viewmodel/checkout.viewmodel';
 import { Checkout } from '@prisma/client';
 import { AddressViewmodel } from '../client/address/viewmodel/address.viewmodel';
 import { CartService } from '../cart/cart.service';
 import { DelivererViewmodel } from '../product/deliverer/viewmodel';
+import { SetCheckoutPaymentMethodDTO } from './dto';
 
 @Injectable()
 export class CheckoutService {
@@ -171,4 +172,117 @@ export class CheckoutService {
 
         return await this.fillCheckoutWithData(checkout)
     }
+
+    async setCheckoutAddress(clientId: string, checkoutId: string, addressId: string): Promise<null> {
+        const c = await this.db.checkout.findUnique({
+            where: { id: checkoutId },
+        });
+        if (!c) {
+            throw new NotFoundException('Checkout não encontrado');
+        }
+        const a = await this.db.address.findUnique({
+            where: { id: addressId },
+        });
+        if (!a) {
+            throw new NotFoundException('Endereço não encontrado');
+        }
+        const cart = await this.db.cart.findMany({
+            where: {
+                clientId: clientId,
+                status: "OPEN"
+            }
+        })
+        if (cart.length > 1) {
+            throw new NotFoundException('Mais de um carrinho aberto');
+        }
+        const validCart = cart[0].id == c.cartId
+        if (!validCart) {
+            throw new NotFoundException('Carrinho não encontrado');
+        }
+
+        await this.db.checkout.update({
+            where: { id: checkoutId },
+            data: {
+                addressId: addressId,
+                status: "ENTREGADOR_PENDENTE",
+            },
+        });
+
+        return null;
+    }
+
+    async setCheckoutDeliverer(clientId: string, checkoutId: string, delivererId: string): Promise<null> {
+        const c = await this.db.checkout.findUnique({
+            where: { id: checkoutId },
+        });
+        if (!c) {
+            throw new NotFoundException('Checkout não encontrado');
+        }
+        const a = await this.db.deliverer.findUnique({
+            where: { id: delivererId },
+        });
+        if (!a) {
+            throw new NotFoundException('Entregador não encontrado');
+        }
+        const cart = await this.db.cart.findMany({
+            where: {
+                clientId: clientId,
+                status: "OPEN"
+            }
+        })
+        if (cart.length > 1) {
+            throw new NotFoundException('Mais de um carrinho aberto');
+        }
+        const validCart = cart[0].id == c.cartId
+        if (!validCart) {
+            throw new NotFoundException('Carrinho não encontrado');
+        }
+
+        await this.db.checkout.update({
+            where: { id: checkoutId },
+            data: {
+                delivererId: delivererId,
+                status: "METODO_DE_PAGAMENTO_PENDENTE"
+            },
+        });
+
+        return null;
+    }
+
+    async setCheckoutPaymentMethod(clientId: string, checkoutId: string, dto: SetCheckoutPaymentMethodDTO): Promise<null> {
+        const c = await this.db.client.findUnique({ where: { id: clientId } })
+        if (!c) {
+            throw new NotFoundException('Cliente não encontrado')
+        }
+        const checkout = await this.db.checkout.findUnique({ where: { id: checkoutId } })
+        if (!checkout) {
+            throw new NotFoundException('Checkout não encontrado')
+        }
+
+        if (dto.paymentMethod === 'credit-card') {
+            await this.handleCreditCardPaymentMethod(clientId, dto.methodId)
+        } else if (dto.paymentMethod === 'bank-slip') {
+            await this.handleBankSlipPaymentMethod(clientId, dto.methodId)
+        } else {
+            throw new BadRequestException('Método de pagamento inválido')
+        }
+        return
+    }
+
+    // WORK IN PROGRESS HERE
+    private async handleCreditCardPaymentMethod(checkoutId: string, id: string) {
+        const creditCard = await this.db.creditCard.findUnique({ where: { id: id } })
+        if (!creditCard) {
+            throw new NotFoundException('Cartão de crédito não encontrado')
+        }
+
+        // WIP
+        // await this.db.checkout.update({
+        //     where: {
+        //         id: checkoutId,
+
+        //     }
+        // })
+    }
+    private async handleBankSlipPaymentMethod(checkoutId: string, id: string) { }
 }
