+++ api/src/checkout/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './checkout-payment.dto';
