--- api/src/checkout/payment/dto/index.ts
@@ -1,1 +1,0 @@
-export * from './checkout-payment.dto'
