--- api/src/checkout/payment/payment.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { PaymentController } from './payment.controller';
-
-describe('PaymentController', () => {
-  let controller: PaymentController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [PaymentController],
-    }).compile();
-
-    controller = module.get<PaymentController>(PaymentController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
